Configuração do eslint + prettier, e modificação na estrutura de pastas
