--- pages/index.js
@@ -1,16 +1,0 @@
-export default function Home() {
-  return (
-    <>
-      <h1>Página principal</h1>
-      <a href="/busca">
-        <h2>Ir para busca</h2>
-      </a>
-      <a href="/sobre">
-        <h2>Sobre nós</h2>
-      </a>
-      <a href="/colabore">
-        <h2>Colabore conosco</h2>
-      </a>
-    </>
-  );
-}
