--- pages/sobre.js
@@ -1,10 +1,0 @@
-export default function About() {
-  return (
-    <>
-      <h1>Página sobre nós</h1>
-      <a href="/">
-        <h2>Voltar para Home</h2>
-      </a>
-    </>
-  );
-}
