+++ src/pages/_app.js
@@ -1,0 +1,9 @@
+import '../styles/globals.scss';
+
+// eslint-disable-next-line react/prop-types
+function MyApp({ Component, pageProps }) {
+  // eslint-disable-next-line react/jsx-props-no-spreading
+  return <Component {...pageProps} />;
+}
+
+export default MyApp;
