+++ src/pages/busca.js
@@ -1,0 +1,10 @@
+export default function Search() {
+  return (
+    <>
+      <h1>Página de Busca</h1>
+      <a href="/">
+        <h2>Voltar para Home</h2>
+      </a>
+    </>
+  );
+}
