+++ src/pages/colabore.js
@@ -1,0 +1,10 @@
+export default function About() {
+  return (
+    <>
+      <h1>Página de Colaboração</h1>
+      <a href="/">
+        <h2>Voltar para Home</h2>
+      </a>
+    </>
+  );
+}
